+++ tests/api_resources/test_search.py
@@ -38,7 +38,9 @@ class TestSearch:
                     "currency": "USD",
                     "keyword_search_only": True,
                     "language": "en",
+                    "length_unit": "mm",
                     "mode": "keyword",
+                    "weight_unit": "mg",
                 },
                 filters={
                     "age": ["newborn"],
@@ -55,6 +57,28 @@ class TestSearch:
                         ]
                     },
                     "condition": "new",
+                    "dimensions": {
+                        "height": {
+                            "unit": "mm",
+                            "max": 0,
+                            "min": 0,
+                        },
+                        "length": {
+                            "unit": "mm",
+                            "max": 0,
+                            "min": 0,
+                        },
+                        "weight": {
+                            "unit": "mg",
+                            "max": 0,
+                            "min": 0,
+                        },
+                        "width": {
+                            "unit": "mm",
+                            "max": 0,
+                            "min": 0,
+                        },
+                    },
                     "exclude_brand_ids": ["string"],
                     "exclude_category_ids": ["string"],
                     "exclude_website_ids": ["string"],
@@ -123,7 +147,9 @@ class TestAsyncSearch:
                     "currency": "USD",
                     "keyword_search_only": True,
                     "language": "en",
+                    "length_unit": "mm",
                     "mode": "keyword",
+                    "weight_unit": "mg",
                 },
                 filters={
                     "age": ["newborn"],
@@ -140,6 +166,28 @@ class TestAsyncSearch:
                         ]
                     },
                     "condition": "new",
+                    "dimensions": {
+                        "height": {
+                            "unit": "mm",
+                            "max": 0,
+                            "min": 0,
+                        },
+                        "length": {
+                            "unit": "mm",
+                            "max": 0,
+                            "min": 0,
+                        },
+                        "weight": {
+                            "unit": "mg",
+                            "max": 0,
+                            "min": 0,
+                        },
+                        "width": {
+                            "unit": "mm",
+                            "max": 0,
+                            "min": 0,
+                        },
+                    },
                     "exclude_brand_ids": ["string"],
                     "exclude_category_ids": ["string"],
                     "exclude_website_ids": ["string"],